Refactored project (#8)
