+++ requirements.txt
@@ -8,6 +8,8 @@ torchvision==0.22.0
 scikit-learn==1.6.1
 ultralytics==8.3.74
 pyyaml==6.0.2
+sphinx==8.2.3
+sphinx-rtd-theme==3.0.2
 
 ruff==0.6.4
 pre-commit==3.8.0
