+++ src/domain/services/image_converter.py
