+++ src/infrastructure/data_loaders.py
@@ -2,7 +2,7 @@ from pathlib import Path
 from typing import List
 import pandas as pd
 
-from domain.camera import CameraDataFromCsv
+from src.domain.camera import CameraDataFromCsv
 
 
 class CameraDataLoader:
